* FEATURE: Updated containers to include sets/reads that use global settings, rather than always requiring fresh settings instances for each set/read * BUGFIX: Fixed warnings in test_fragmentation and test_karl_containers * BUGFIX: Fixed Map container to use supplied settings instead of ony using global settings * Updated version 1.7.6
git-svn-id: http://madara.googlecode.com/svn/trunk@1325 2850746a-f1c8-11dd-b507-8b4ecd5210be

+++ include/madara/knowledge_engine/containers/Reliable_File.cpp
@@ -170,21 +170,22 @@ Madara::Knowledge_Engine::Containers::Reliable_File::split (void)
   if (size_ % max_frag_size_ != 0)
   {
     extra = true;
-    fragments_.resize (frags);
+    fragments_.resize ((int)frags);
   }
   else
   {
-    fragments_.resize (frags + 1);
+    fragments_.resize ((int)(frags + 1));
   }
 
   for (; i < frags; ++i, buffer += max_frag_size_)
   {
-    fragments_.set_file (i, buffer, max_frag_size_, settings);
+    fragments_.set_file ((unsigned int)i, buffer, max_frag_size_, settings);
   }
 
   if (extra)
   {
-    fragments_.set_file (i + 1, buffer, size_ % max_frag_size_, settings);
+    fragments_.set_file ((unsigned int)(i + 1), buffer,
+      size_ % max_frag_size_, settings);
   }
 }
